Bugfix: memory issue when deleting balls
When deleting a ball, if the last ball in the list was in a socket the method didn't wrong since the socket holds a pointer to that memory, which was moved with the old system.
I replaced that by a simpler (and a bit more expensive) model where we check for valid IDs instead of moving memory around.
Also submitting some fixes for the scoreboard which were already in the build.